Implement BigTIFF Necessity Check
Description:
This commit introduces a check to determine whether the `bigtiff=True` parameter should be used when exporting TIFF files (save_movie=True). The necessity of using BigTIFF is assessed based on the size of the data array to be written. If the total size of the array is close to or exceeds 4 GB, `bigtiff=True` is used to prevent errors related to file size limitations in the standard TIFF format. This enhancement aims to avoid the "data too large for non-BigTIFF file" error that occurs when attempting to write large files in the standard TIFF format.